refactor: 수동 로또의 개수를 검증 및 생성하는 것을 GenerateCount의 책임으로 이동에 따른 함수 및 테스트 코드 수정
